+++ custom_components/breakage_radar/sensor.py
@@ -70,6 +70,7 @@ class BreakageRadarSensor(CoordinatorEntity[BreakageRadarCoordinator], SensorEnt
             ATTR_DETAILS: data.get("details", []),
             ATTR_INDEX_GENERATED: data.get("index_generated_utc", ""),
             "affected_domains": data.get("affected_domains", []),
+            "schedule": data.get("schedule", []),
             "broken_now": data.get("broken_now", {}),
             "broken_now_count": data.get("broken_now_count", 0),
             "imminent": data.get("imminent", {}),